refactor: 타입 모듈 import 시 type 키워드 추가 (#78)
