+++ src/shared/ui/atoms/avatar/foundation/styles.ts
