+++ src/shared/ui/atoms/button/foundation/index.ts
@@ -1,2 +1,2 @@
-export { default as BaseButton, Props as BaseButtonProps } from "./BsaeButton";
+export { default as BaseButton, type Props as BaseButtonProps } from "./BsaeButton";
 export { type ButtonContainerPalette } from "./styles";
