+++ src/shared/ui/atoms/button/variant/BothButton.tsx
@@ -1,8 +1,8 @@
 import clsx from "clsx";
-import { TypographyKey } from "@shared/tokens/typography";
+import type { TypographyKey } from "@shared/tokens/typography";
 import { Icon, Text } from "@shared/ui/atoms";
 import type { IconName } from "@shared/ui/atoms/icon/variants";
-import { BaseButton, BaseButtonProps, ButtonContainerPalette } from "../foundation";
+import { BaseButton, type BaseButtonProps, type ButtonContainerPalette } from "../foundation";
 
 interface Palette extends ButtonContainerPalette {
   textColor: string;
