+++ src/shared/ui/atoms/floating-button/variant/index.ts
@@ -1,2 +1,2 @@
-export { default as IconFloatingButton, Props as IconFloatingButtonProps } from "./IconFloatingButton";
-export { default as ExtendedFloatingButton, Props as ExtendedFloatingButtonProps } from "./ExtendedFloatingButton";
+export { default as IconFloatingButton, type Props as IconFloatingButtonProps } from "./IconFloatingButton";
+export { default as ExtendedFloatingButton, type Props as ExtendedFloatingButtonProps } from "./ExtendedFloatingButton";
