+++ src/shared/ui/atoms/input/variant/SearchInput.tsx
@@ -1,5 +1,5 @@
-import { TextInput, TextInputProps, View } from "react-native";
-import { InputRequiredProps, InputStyle } from "../foundation";
+import { TextInput, type TextInputProps, View } from "react-native";
+import { type InputRequiredProps, InputStyle } from "../foundation";
 import ClearButton from "../part/ClearButton";
 
 export type Props = InputRequiredProps & Omit<TextInputProps, keyof InputRequiredProps>;
