+++ src/shared/ui/atoms/input/variant/UnderlineInput.tsx
@@ -1,8 +1,8 @@
 import clsx from "clsx";
-import { Pressable, TextInput, TextInputProps, View } from "react-native";
+import { Pressable, TextInput, type TextInputProps, View } from "react-native";
 import { Icon } from "@shared/ui/atoms";
 import type { IconName } from "@shared/ui/atoms/icon/variants";
-import { InputRequiredProps, InputStyle } from "../foundation";
+import { type InputRequiredProps, InputStyle } from "../foundation";
 import ClearButton from "../part/ClearButton";
 
 export interface Props extends InputRequiredProps, Omit<TextInputProps, keyof InputRequiredProps> {
