+++ src/shared/ui/atoms/text/index.tsx
@@ -1,5 +1,5 @@
 import clsx from "clsx";
-import { Text as RNText, TextProps } from "react-native";
+import { Text as RNText, type TextProps } from "react-native";
 import { type TypographyKey, typographyTokens } from "@shared/tokens/typography";
 
 interface Props extends TextProps {
